requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Retry logic
tenacity>=8.2.0

# Web scraping
beautifulsoup4>=4.12.0

# Testing
pytest>=7.4.0
httpx>=0.25.0
